pyproject.toml
[project]
name = "agentx-qa-evaluator"
version = "0.1.0"
description = "Dialectical-TAS-Evaluator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.10.0",
    "uvicorn>=0.30.0",
    "rapidfuzz>=3.6.1",
    "httpx>=0.26.0",
    "numpy>=1.26.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]